src/topdup_open/requirements.txt
aniso8601==8.0.0
astroid==2.4.2
autopep8==1.5.4
cachetools==4.1.1
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
importlib-metadata==1.7.0
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
jsonpickle==1.4.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.2
pandas==1.1.2
pika==1.1.0
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.19
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.49.0
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
